reporter + test "light" or "full" + report "light" or "full"
