+++ config/default/default_config.js
@@ -1,12 +1,13 @@
 module.exports = {
     features_path: "./features",
     steps_path: "./steps/common-steps",
+    mode: 'full', //'full' or 'light'
     components: {
         enabled: false,
         components_path: "./components/components",
         excludedComponents: []
     },
     reporter: {
-        type: 'light' //'full' or 'light'
+        type: 'full' //'full' or 'light'
     }
 }
