+++ lib/app.js
@@ -1,11 +1,9 @@
 let tape = require('tape');
 let path = require('path');
 let utils = require('./utils/utilities');
-let stepsCache = [];
-let reporter = require('../reporter/sources/index')
 
-let checkComponent = require('./test-core/checkComponent');
-let checkStepMatch = require('./test-core/checkStepMatch');
+let reporter = require('../reporter/sources/index')
+let testRunner = require('./test-core/test-runner')
 
 function GherkinChecker(cg_config) {
     this.config = cg_config;
@@ -63,29 +61,16 @@ GherkinChecker.prototype.run = function () {
     utils.readDirRecurs(path.resolve(self.config.features_path), (err, data) => {
         if (err) throw err;
 
+        //reporter
         tape.createStream()
             .pipe(reporter(self.config.reporter))
             .pipe(process.stdout)
 
-        tape('Check gherkin consistency', assert => {
-            data.forEach(file => {
-                let fileContent = utils.readFileSync(file);
-                if (!fileContent) return;
-                var compiledGherkin = utils.getCompiledGherkin(fileContent);
-                compiledGherkin.forEach(scenario => {
-                    scenario.steps.forEach(step => {
-                        //skip already checked steps
-                        if (stepsCache.indexOf(step.text) > -1) return;
-                        stepsCache.push(step.text);
+        if(self.config.mode === 'light')
+            testRunner.singleTestRun.call(self, tape, data);
+        else if(self.config.mode === 'full')
+            testRunner.fullTestRun.call(self, tape, data);
 
-                        checkComponent.call(self, step.text, assert);
-                        checkStepMatch.call(self, step.text, assert);
-                    })
-                })
-
-            });
-            assert.end();
-        });
     });
 };
