+++ lib/utils/utilities.js
@@ -46,6 +46,10 @@ exports.readFileAsync = (file) => {
     })
 }
 
+exports.getFileName = (file) => {
+    return path.basename(file);
+}
+
 /**
  * read file synchronously
  * @param file
